backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.1
django-cors-headers==4.4.0
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==3.1.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.3
idna==3.8
jwcrypto==1.5.6
mysqlclient==2.2.4
oauthlib==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
flake8==5.0.0